fix: extract sidebar specific code from modal
