+++ projects/angular/src/modal/modal.html
@@ -28,15 +28,7 @@
 
       <div class="modal-content">
         <div class="modal-header--accessible">
-          <button
-            type="button"
-            [attr.aria-label]="commonStrings.keys.sidePanelPin"
-            class="pinnable"
-            *ngIf="pinnable"
-            (click)="togglePinned()"
-          >
-            <cds-icon [attr.shape]="pinned ? 'unpin' : 'pin'"></cds-icon>
-          </button>
+          <ng-content select=".leading-button"></ng-content>
           <div class="modal-title-wrapper" #title id="{{modalId}}" cdkFocusInitial tabindex="-1">
             <ng-content select=".modal-title"></ng-content>
           </div>
