+++ projects/angular/src/modal/side-panel.html
@@ -8,7 +8,17 @@
   [clrModalLabelledById]="labelledById"
   [clrModalPreventClose]="preventClose"
   (clrModalAlternateClose)="altClose.emit($event)"
+  [clrModalOverrideScrollService]="true"
 >
+  <button
+    type="button"
+    [attr.aria-label]="commonStrings.keys.sidePanelPin"
+    class="leading-button pinnable"
+    *ngIf="clrSidePanelPinnable"
+    (click)="togglePinned()"
+  >
+    <cds-icon [attr.shape]="pinned ? 'unpin' : 'pin'"></cds-icon>
+  </button>
   <div class="modal-title"><ng-content select=".side-panel-title"></ng-content></div>
   <div class="modal-body"><ng-content select=".side-panel-body"></ng-content></div>
   <div class="modal-footer"><ng-content select=".side-panel-footer"></ng-content></div>
