+++ projects/angular/src/modal/side-panel.ts
@@ -6,17 +6,20 @@
  */
 
 import {
-  AfterViewInit,
   Component,
   ElementRef,
   EventEmitter,
   HostListener,
   Input,
+  OnChanges,
+  OnDestroy,
   OnInit,
   Output,
+  SimpleChange,
   ViewChild,
 } from '@angular/core';
 
+import { ClrCommonStringsService } from '../utils';
 import { ClrModal } from './modal';
 import { ClrModalConfigurationService } from './modal-configuration.service';
 
@@ -27,22 +30,61 @@ import { ClrModalConfigurationService } from './modal-configuration.service';
     '[class.side-panel]': 'true',
   },
 })
-export class ClrSidePanel implements OnInit, AfterViewInit {
+export class ClrSidePanel implements OnInit, OnDestroy, OnChanges {
   @Input('clrSidePanelOpen') _open = false;
   @Output('clrSidePanelOpenChange') openChange = new EventEmitter<boolean>(false);
   @Input('clrSidePanelCloseButtonAriaLabel') closeButtonAriaLabel: string | undefined;
-  @Input('clrSidePanelSize') size: string;
   @Input('clrSidePanelSkipAnimation') skipAnimation = false;
   @Input('clrSidePanelLabelledById') labelledById: string;
   @Input('clrSidePanelStaticBackdrop') staticBackdrop = false;
   @Input('clrSidePanelPreventClose') preventClose = false;
   @Output('clrSidePanelAlternateClose') altClose = new EventEmitter<boolean>(false);
-
   @ViewChild(ClrModal) private modal: ClrModal;
 
   private _pinnable = false;
+  private _pinned = false;
+  private originalStopClose: boolean;
 
-  constructor(private element: ElementRef<HTMLElement>, private configuration: ClrModalConfigurationService) {}
+  private _size: string;
+
+  constructor(
+    private element: ElementRef<HTMLElement>,
+    private configuration: ClrModalConfigurationService,
+    public commonStrings: ClrCommonStringsService
+  ) {}
+
+  @Input('clrSidePanelSize')
+  get size(): string {
+    return this._size;
+  }
+
+  set size(value: string) {
+    if (this._size !== value) {
+      this._size = value;
+      if (this.clrSidePanelPinnable && this.pinned) {
+        this.displayOverlapping();
+        this.displaySideBySide();
+      }
+    }
+  }
+
+  get pinned(): boolean {
+    return this._pinned;
+  }
+
+  set pinned(pinned: boolean) {
+    if (this.clrSidePanelPinnable) {
+      this._pinned = pinned;
+      if (pinned) {
+        this.originalStopClose = this.modal.stopClose;
+        this.modal.stopClose = true;
+        this.displaySideBySide();
+      } else {
+        this.modal.stopClose = this.originalStopClose;
+        this.displayOverlapping();
+      }
+    }
+  }
 
   @Input()
   get clrSidePanelBackdrop(): boolean {
@@ -62,17 +104,32 @@ export class ClrSidePanel implements OnInit, AfterViewInit {
 
   set clrSidePanelPinnable(pinnable: boolean) {
     this._pinnable = pinnable;
-    if (this.modal) {
-      this.modal.pinnable = pinnable;
-    }
+  }
+
+  private get hostElement(): HTMLElement {
+    return (this.element.nativeElement as HTMLElement).closest('.clr-modal-host') || document.body;
   }
 
   ngOnInit(): void {
     this.configuration.fadeMove = 'fadeLeft';
   }
 
-  ngAfterViewInit() {
-    this.modal.pinnable = this._pinnable;
+  ngOnChanges(changes: { [propName: string]: SimpleChange }): void {
+    if (changes && Object.prototype.hasOwnProperty.call(changes, '_open')) {
+      if (changes._open.currentValue) {
+        if (this.clrSidePanelPinnable && this.pinned) {
+          this.displaySideBySide();
+        }
+      } else {
+        if (this.clrSidePanelPinnable && this.pinned) {
+          this.displayOverlapping();
+        }
+      }
+    }
+  }
+
+  ngOnDestroy(): void {
+    this.displayOverlapping();
   }
 
   open() {
@@ -81,6 +138,10 @@ export class ClrSidePanel implements OnInit, AfterViewInit {
 
   close() {
     this.modal.close();
+  }
+
+  togglePinned() {
+    this.pinned = !this.pinned;
   }
 
   @HostListener('document:pointerup', ['$event'])
@@ -92,5 +153,17 @@ export class ClrSidePanel implements OnInit, AfterViewInit {
     ) {
       this.modal.close();
     }
+  }
+
+  private displaySideBySide() {
+    this.hostElement.classList.add('clr-side-panel-pinned-' + this.size);
+  }
+
+  private displayOverlapping() {
+    this.hostElement.classList.forEach(className => {
+      if (className.startsWith('clr-side-panel-pinned-')) {
+        this.hostElement.classList.remove(className);
+      }
+    });
   }
 }
